ref(flutter): Group app-start by standalone vs ui.load
Move app-start modules under standalone/ and ui_load_attached/,
lift shared data to app_start/, and dissolve app_start_constants into
the types that own those values.
