--- packages/flutter/lib/src/app_start/app_start_constants.dart
@@ -1,20 +1,0 @@
-import 'package:meta/meta.dart';
-
-@internal
-const appStartRootName = 'App Start';
-
-@internal
-const appStartIdleTimeout = Duration(seconds: 3);
-
-@internal
-const appStartFinalTimeout = Duration(seconds: 30);
-
-@internal
-const appStartPluginRegistrationDescription =
-    'App start to plugin registration';
-
-@internal
-const appStartSentrySetupDescription = 'Before Sentry Init Setup';
-
-@internal
-const appStartFirstFrameRenderDescription = 'First frame render';
