--- packages/flutter/lib/src/app_start/app_start_trace.dart
@@ -1,13 +1,0 @@
-import 'dart:async';
-
-import 'package:meta/meta.dart';
-
-/// Lifecycle-independent operations for a standalone app-start trace.
-@internal
-abstract interface class AppStartTrace {
-  void recordFirstFrame(DateTime endTimestamp);
-
-  void finish(DateTime endTimestamp);
-
-  FutureOr<void> close();
-}
