+++ packages/flutter/lib/src/app_start/standalone/app_start_trace.dart
@@ -1,0 +1,25 @@
+import 'dart:async';
+
+import 'package:meta/meta.dart';
+
+/// Standalone root transaction / segment name.
+@internal
+const standaloneAppStartRootName = 'App Start';
+
+/// Idle auto-finish after the natural first-frame end.
+@internal
+const standaloneAppStartIdleTimeout = Duration(seconds: 3);
+
+/// Hard deadline from trace creation when first frame never arrives.
+@internal
+const standaloneAppStartFinalTimeout = Duration(seconds: 30);
+
+/// Lifecycle-independent operations for a standalone app-start trace.
+@internal
+abstract interface class AppStartTrace {
+  void recordFirstFrame(DateTime endTimestamp);
+
+  void finish(DateTime endTimestamp);
+
+  FutureOr<void> close();
+}
