+++ packages/flutter/lib/src/app_start/ui_load_attached/native_app_start_handler_v2.dart
@@ -2,11 +2,10 @@
 
 import 'package:meta/meta.dart';
 
-import '../../sentry_flutter.dart';
-import '../app_start/app_start_constants.dart';
-import '../app_start/app_start_data.dart';
-import '../native/sentry_native_binding.dart';
-import '../utils/internal_logger.dart';
+import '../../../sentry_flutter.dart';
+import '../app_start_data.dart';
+import '../../native/sentry_native_binding.dart';
+import '../../utils/internal_logger.dart';
 
 /// V2 handler for native app start spans using the streaming span API.
 @internal
