+++ packages/flutter/lib/src/sentry_flutter.dart
@@ -17,17 +17,17 @@ import 'event_processor/widget_event_processor.dart';
 import 'file_system_transport.dart';
 import 'flutter_exception_type_identifier.dart';
 import 'frame_callback_handler.dart';
-import 'app_start/standalone_app_start_integration.dart';
-import 'app_start/standalone_app_start_lifecycle.dart';
+import 'app_start/standalone/standalone_app_start_integration.dart';
+import 'app_start/standalone/standalone_app_start_lifecycle.dart';
+import 'app_start/ui_load_attached/generic_app_start_integration.dart';
+import 'app_start/ui_load_attached/native_app_start_handler.dart';
+import 'app_start/ui_load_attached/native_app_start_handler_v2.dart';
 import 'integrations/connectivity/connectivity_integration.dart';
 import 'integrations/flutter_framework_feature_flag_integration.dart';
 import 'integrations/frames_tracking_integration.dart';
 import 'integrations/integrations.dart';
-import 'integrations/native_app_start_handler.dart';
-import 'integrations/native_app_start_handler_v2.dart';
 import 'integrations/replay_telemetry_integration.dart';
 import 'integrations/screenshot_integration.dart';
-import 'app_start/generic_app_start_integration.dart';
 import 'integrations/native_trace_sync_integration.dart';
 import 'integrations/thread_info_integration.dart';
 import 'integrations/web_session_integration.dart';
