+++ packages/flutter/test/app_start/standalone/static_app_start_trace_test.dart
@@ -3,9 +3,9 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
 import 'package:sentry_flutter/src/app_start/app_start_data.dart';
-import 'package:sentry_flutter/src/app_start/static_app_start_trace.dart';
+import 'package:sentry_flutter/src/app_start/standalone/static_app_start_trace.dart';
 
-import '../mocks.dart';
+import '../../mocks.dart';
 
 void main() {
   group('$StaticAppStartTrace', () {
